<refactor> Modify code name to look intuitive(pomodoro_service -> pomodoro_timer_service)
