+++ lib/screens/timer_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../services/pomodoro_service.dart';
+import '../services/pomodoro_timer_service.dart';
 import '../services/stopwatch_service.dart';
 import '../components/timer/timer_pomodoro.dart';
 import '../components/timer/timer_stopwatch.dart'; // StopwatchPage import 추가
